rewrite cl files

+++ src/kernels/intersect.cl
@@ -1,3 +1,5 @@
+///////
+
 #include "intersect.h"
 
 float intersectPlane(struct Ray ray, struct Plane plane)
@@ -6,7 +8,7 @@ float intersectPlane(struct Ray ray, struct Plane plane)
     if (d != 0.0f)
         return -(dot(plane.normal, ray.origin) + plane.offset) / d;
     else
-        return INFINITY;
+        return (float)INFINITY;
 }
 
 float intersectSphere(struct Ray ray, struct Sphere sphere)
@@ -15,14 +17,14 @@ float intersectSphere(struct Ray ray, struct Sphere sphere)
     float b = dot(ray.direction, oc);
     float c = dot(oc, oc) - pow(sphere.radius, 2.0f);
     float d = pow(b, 2.0f) - c;
-
+    
     if (d < 0.0f)
-        return INFINITY;
+        return (float)INFINITY;
     else
         return -b - sqrt(d);
 }
 
 float intersectTriangle(struct Ray ray, struct Triangle triangle)
 {
-    return INFINITY;
+    return (float)INFINITY;
 }
